third_party/blink/web_tests/external/wpt/html/semantics/scripting-1/the-script-element/module/charset-02-expected.txt

This is a testharness.js-based test.
FAIL UTF-8 module script with UTF-8 BOM assert_equals: Should be decoded as UTF-8 expected (string) "三村かな子" but got (undefined) undefined
PASS UTF-16 module script with UTF-16BE BOM
PASS UTF-16 module script with UTF-16LE BOM
Harness: the test ran to completion.
